move dashboardConfig out of config
