+++ frontend/src/dashboardConfig/dashboardConfigQueryError.test.ts
