+++ frontend/src/dashboardConfig/dashboardEnums.ts
