+++ frontend/src/dashboardConfig/fetchDashboardConfig.ts
