+++ frontend/src/dashboardConfig/schema.test.ts
