+++ frontend/vite.config.ts
@@ -101,6 +101,7 @@ export default defineConfig({
       components: '/src/components',
       config: '/src/config',
       context: '/src/context',
+      dashboardConfig: '/src/dashboardConfig',
       hooks: '/src/hooks',
       utils: '/src/utils',
       muiTheme: '/src/muiTheme',
